Add files via upload
This is a simple web app that allows users to count up or down using buttons on the screen. The app is built using HTML, CSS, and JavaScript, and is designed to be easy to use and customize. Features include:
* A counter that displays the current count on the screen
* Buttons to increment or decrement the count
* A button to reset the count to zero
* Clean and simple UI with modern design
This project is a great starting point for those new to web development or looking to build simple web apps. It is fully responsive and can be used on desktop or mobile devices. Feel free to fork or clone this repository to build your own version of the app or to use it as a template for your own projects